+++ src/epeios/CHANGELOG.md
@@ -5,6 +5,11 @@
 - [adpt]: Adaptation to changes in underlying modules;
 - [stdz]: modification to comply with current standards;
 
+## 2018-09-09
+
+- MTX:
+  - adding `rMutex::InitAndLock(...)`,
+
 ## 2018-08-16
 
 - CSDBNS: